+++ PenyFront/src/shared/components/GlobalSearch.tsx
@@ -6,18 +6,16 @@ import {
     Text,
     Group,
     Badge,
-    Avatar,
     Loader,
     Box,
     ScrollArea,
     ActionIcon,
-    Divider,
-    UnstyledButton,
     CloseButton,
 } from '@mantine/core';
 import { useDebouncedValue } from '@mantine/hooks';
-import { Search, X, Calendar, FileText, MapPin } from 'lucide-react';
+import { Search, X } from 'lucide-react';
 import { prisonersService } from '../services/prisonersService';
+import { SearchResultItem } from './SearchResultItem';
 import type { PrisionerListItem } from '../types/prisonerTypes';
 
 interface GlobalSearchProps {
@@ -28,7 +26,7 @@ interface GlobalSearchProps {
 
 export function GlobalSearch({ opened, onClose, onSelectResult }: GlobalSearchProps) {
     const [searchQuery, setSearchQuery] = useState('');
-    const [debouncedQuery] = useDebouncedValue(searchQuery, 300);
+    const [debouncedQuery] = useDebouncedValue(searchQuery, 150);
     const [results, setResults] = useState<PrisionerListItem[]>([]);
     const [loading, setLoading] = useState(false);
     const [total, setTotal] = useState(0);
@@ -90,24 +88,6 @@ export function GlobalSearch({ opened, onClose, onSelectResult }: GlobalSearchPr
         handleClose();
     };
 
-    const getStatusColor = (status: string) => {
-        const colors: Record<string, string> = {
-            Activo: 'green',
-            Trasladado: 'blue',
-            Liberado: 'gray',
-            Archivado: 'red',
-        };
-        return colors[status] || 'gray';
-    };
-
-    const formatDate = (dateString: string) => {
-        return new Date(dateString).toLocaleDateString('es-ES', {
-            year: 'numeric',
-            month: 'short',
-            day: 'numeric',
-        });
-    };
-
     return (
         <Modal
             yOffset="10vh"
@@ -115,12 +95,11 @@ export function GlobalSearch({ opened, onClose, onSelectResult }: GlobalSearchPr
             onClose={handleClose}
             size="lg"
             padding={0}
-
             withCloseButton={false}
             overlayProps={{
                 opacity: 0.55,
-                blur: 8,
             }}
+            transitionProps={{ transition: 'fade', duration: 150 }}
             styles={{
                 body: { padding: 0 },
                 content: { overflow: 'hidden' },
@@ -183,117 +162,12 @@ export function GlobalSearch({ opened, onClose, onSelectResult }: GlobalSearchPr
                             </Box>
                             <Stack gap={0}>
                                 {results.map((item, index) => (
-                                    <Box key={item.prisoner.id}>
-                                        <UnstyledButton
-                                            onClick={() => handleSelectResult(item)}
-                                            style={{
-                                                width: '100%',
-                                                padding: 'var(--mantine-spacing-md)',
-                                                transition: 'background-color 0.2s',
-                                            }}
-                                            onMouseEnter={(e) => {
-                                                e.currentTarget.style.backgroundColor = 'var(--mantine-color-gray-0)';
-                                            }}
-                                            onMouseLeave={(e) => {
-                                                e.currentTarget.style.backgroundColor = 'transparent';
-                                            }}
-                                        >
-                                            <Group wrap="nowrap" gap="md">
-                                                <div className='flex flex-row justify-between w-full'>
-                                                    <Box style={{ flex: 1, minWidth: 0 }}>
-                                                        <Group gap="xs" mb={4}>
-                                                            <Text fw={500} size="sm" lineClamp={1}>
-                                                                {item.identity?.first_name} {item.identity?.surname}
-                                                            </Text>
-                                                            <Badge
-                                                                size="xs"
-                                                                color={getStatusColor(item.prisoner.status)}
-                                                                variant="light"
-                                                            >
-                                                                {item.prisoner.status}
-                                                            </Badge>
-                                                            <Badge size="xs" color="gray" variant="light">
-                                                                {item.penitentiary?.category ?? 'Sin categoría'}
-                                                            </Badge>
-                                                        </Group>
-                                                        <div className='flex flex-row justify-start w-full '>
-                                                            {/* Columna datos personales */}
-                                                            <Stack gap={4}>
-                                                                <Group gap="xs">
-                                                                    <FileText size={12} style={{ color: 'var(--mantine-color-dimmed)' }} />
-                                                                    <Text size="xs" c="dimmed">
-                                                                        Registro: {item.prisoner.registration_number}
-                                                                    </Text>
-                                                                </Group>
-
-                                                                <Group gap="xs">
-                                                                    <Calendar size={12} style={{ color: 'var(--mantine-color-dimmed)' }} />
-                                                                    <Text size="xs" c="dimmed">
-                                                                        Ingreso: {formatDate(item.prisoner.admission_date)}
-                                                                    </Text>
-                                                                </Group>
-
-                                                                {item.identity?.residence && (
-                                                                    <Group gap="xs">
-                                                                        <MapPin size={12} style={{ color: 'var(--mantine-color-dimmed)' }} />
-                                                                        <Text size="xs" c="dimmed" lineClamp={1}>
-                                                                            {item.identity?.residence}
-                                                                        </Text>
-                                                                    </Group>
-                                                                )}
-                                                                {item.cases?.length > 0 ? (
-                                                                    
-                                                                    <Group gap={4} wrap="wrap">
-                                                                         <Badge size="xs" variant="dot" color="orange">
-                                                                        {item.cases?.length} caso{item.cases?.length !== 1 ? 's' : ''}
-                                                                    </Badge>
-                                                                        {item.cases.map((c, idx) => (
-                                                                            <Badge key={c.id || idx} size="xs" color="orange" variant="light">
-                                                                                {c.crime}
-                                                                            </Badge>
-                                                                        ))}
-                                                                    </Group>
-                                                                ) : (
-                                                                    <Text size="xs" c="dimmed">Sin delitos</Text>
-                                                                )}
-                                                            </Stack>
-                                                            {/* Columna información penitenciaria */}
-                                                            <Stack gap={4} style={{ minWidth: 120 }}>
-
-                                                                {item.penitentiary?.building_number && (
-                                                                    <Text size="xs" c="dimmed">
-                                                                        Edificio: <b>{item.penitentiary.building_number}</b>
-                                                                    </Text>
-                                                                )}
-                                                                {item.penitentiary?.cell_number && (
-                                                                    <Text size="xs" c="dimmed">
-                                                                        Celda: <b>{item.penitentiary.cell_number}</b>
-                                                                    </Text>
-                                                                )}
-                                                                {item.penitentiary?.bed_number && (
-                                                                    <Text size="xs" c="dimmed">
-                                                                        Cama: <b>{item.penitentiary.bed_number}</b>
-                                                                    </Text>
-                                                                )}
-                                                                {!(item.penitentiary?.category || item.penitentiary?.building_number || item.penitentiary?.cell_number || item.penitentiary?.bed_number) && (
-                                                                    <Text size="xs" c="dimmed">Sin datos</Text>
-                                                                )}
-                                                            </Stack>
-
-                                                        </div>
-
-                                                    </Box>
-                                                    <Avatar
-                                                        src={item.identity?.photo_file?.url ?? ''}
-                                                        size={86}
-                                                        radius="md"
-                                                    >
-                                                    </Avatar>
-                                                </div>
-                                            </Group>
-                                        </UnstyledButton>
-                                        {index < results.length - 1 && <Divider />}
-                                    </Box>
+                                    <SearchResultItem
+                                        key={item.prisoner.id}
+                                        item={item}
+                                        isLast={index === results.length - 1}
+                                        onSelect={handleSelectResult}
+                                    />
                                 ))}
                             </Stack>
                         </>
